% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_helpers.R
\name{apply_scaling}
\alias{apply_scaling}
\title{Apply suggested scaling factors to data}
\usage{
apply_scaling(data, scaling_df)
}
\arguments{
\item{data}{A data.frame or matrix containing the data to be scaled.
Column names should match the 'Variable' column in `scaling_df`.}

\item{scaling_df}{A data.frame produced by \code{\link{suggest_scaling}},
containing the instructions on how to scale each variable.}
}
\value{
A new data.frame or matrix with the data scaled according to the
provided instructions.
}
\description{
Applies the scaling operations suggested by \code{\link{suggest_scaling}}
to a dataset.
}
\examples{
# For a complete example, please see the documentation for `suggest_scaling`.
# ?suggest_scaling
}
