% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_helpers.R
\name{ell_features}
\alias{ell_features}
\title{Calculate ellipsoid characteristics (semi-axes, volume, axes coordinates)}
\usage{
ell_features(centroid, covariance_matrix, level = 0.95)
}
\arguments{
\item{centroid}{(numeric) vector of values representing the centroid of the
data of interest.}

\item{covariance_matrix}{numeric matrix containing values of variance and
covariance derived from the data of interest, that define the shape and limits
of the ellipsoid.}

\item{level}{(numeric) the confidence level of a confidence region
for the ellipsoid. Default = 0.95. See details.}
}
\value{
A list containing the data used and the volume, semi-axes length, and axes
coordinates for the ellipsoid described by the arguments defined.
}
\description{
Calculate ellipsoid characteristics (semi-axes, volume, axes coordinates)
}
