% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ell_predict.R
\name{ell_predict}
\alias{ell_predict}
\title{Predict mahalanobis and suitability from ellipsoidal virtual niche}
\usage{
ell_predict(data, features, longitude = NULL, latitude = NULL,
            tol = 1e-60, include_suitability = TRUE,
            include_truncated = TRUE)
}
\arguments{
\item{data}{matrix, data.frame, or SpatRaster containing variables of
interest included in ellipsoid \code{features}.}

\item{features}{list of characteristics that defined the ellipsoid. This
list ca be obtained using the function \code{\link{ell_features}}.}

\item{longitude}{(character) if data is a data.frame or matrix, name of the
column with longitudes (if present). Default = NULL.}

\item{latitude}{(character) if data is a data.frame or matrix, name of the
column with latitudes (if present). Default = NULL.}

\item{tol}{(numeric) the tolerance for detecting linear dependencies.
Default = 1e-8.}

\item{include_suitability}{(logical) whether to include suitability
predictions. Default = TRUE.}

\item{include_truncated}{(logical) whether to include suitability
predictions truncated based on ellipsoid limits (level). Default = TRUE.}
}
\value{
A list of predictions in vector or SpatRaster format, based on the features
of the ellipsoid niche.
}
\description{
Predict mahalanobis and suitability from ellipsoidal virtual niche
}
