% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_helpers.R
\name{ell_volume}
\alias{ell_volume}
\title{Helper function to calculate niche volume}
\usage{
ell_volume(n_dimensions, semi_axes_length)
}
\arguments{
\item{n_dimensions}{(numeric) number of dimensions to be considered.}

\item{semi_axes_length}{(numeric) length of ellipsoid axes.}
}
\value{
A numeric value representing the volume of the ellipsoidal niche.
}
\description{
Helper function to calculate niche volume
}
