% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_helpers.R
\name{export_metadata}
\alias{export_metadata}
\title{Export ellipsoid niche metadata}
\usage{
export_metadata(path, folder_name, parameters, features, overwrite = FALSE)
}
\arguments{
\item{path}{(character) path to the directory where the folder will be created.}

\item{folder_name}{(character) name of the folder to be created to save data.}

\item{parameters}{a data.frame with initial parameters used to create the niche.}

\item{features}{a list of features of the ellipsoid niche obtained with
\code{\link{ell_features}}.}

\item{overwrite}{(logical) whether to overwrite the folder if it already exists.
Default = FALSE.}
}
\value{
Creates a folder containing a .RData file with parameters and
features, and a .txt file with a summary of the metadata.
}
\description{
Saves ellipsoid niche metadata as RData and a TXT file.
}
