% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{plot_2D}
\alias{plot_2D}
\title{Plot 2D ellipsoids in multiple panels}
\usage{
plot_2D(
  features = NULL,
  background = NULL,
  par_list = list(mar = c(4.5, 4.5, 1, 1), asp = NA, cex = 1, col_bg = "black", col_elem
    = "white", y_las = 0),
  lp_list = list(lty = 1, lwd = 1, col_l = "yellow", alpha_l = 1, pch = 1, cex_p = 1,
    col_p = "green", alpha_p = 0.5)
)
}
\arguments{
\item{features}{a list of features of the ellipsoid niche obtained with
\code{\link{ell_features}}. Default = NULL.}

\item{background}{matrix or data.frame of background points to be plotted.
Default = NULL.}

\item{par_list}{a list of parameters for plotting.
\itemize{
  \item \code{mar}: margins of the plot.
  \item \code{asp}: aspect ratio.
  \item \code{cex}: character expansion factor.
  \item \code{col_bg}: background color.
  \item \code{col_elem}: color for plot elements (axes, labels).
  \item \code{y_las}: orientation of y-axis labels.
}}

\item{lp_list}{a list of parameters for lines and points.
\itemize{
  \item \code{lty}: line type.
  \item \code{lwd}: line width.
  \item \code{col_l}: line color.
  \item \code{alpha_l}: alpha for line color.
  \item \code{pch}: point character.
  \item \code{cex_p}: character expansion factor for points.
  \item \code{col_p}: point color.
  \item \code{alpha_p}: alpha for point color.
}}
}
\value{
A 2D plot displaying ellipsoid projections.
}
\description{
Plots all 2D combinations of variables from niche features
or a background dataset.
}
