% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{plot_3D}
\alias{plot_3D}
\title{Plot 3D ellipsoids in multiple panels}
\usage{
plot_3D(
  features = NULL,
  background = NULL,
  par_list = list(asp = c(1, 1, 1), cex = 1, col_bg = "black", col_material = "white",
    box = FALSE),
  lp_list = list(type_ell = "wire", col_ell = "yellow", alpha_ell = 1, size_p = 3, col_p
    = "green", alpha_p = 0.5)
)
}
\arguments{
\item{features}{a list of features of the ellipsoid niche obtained with
\code{\link{ell_features}}. Default = NULL.}

\item{background}{matrix or data.frame of background points to be plotted.
Default = NULL.}

\item{par_list}{a list of parameters for plotting.
\itemize{
  \item \code{asp}: aspect ratio.
  \item \code{cex}: character expansion factor.
  \item \code{col_bg}: background color.
  \item \code{col_material}: color for plot elements (axes, labels).
  \item \code{box}: logical, whether to draw a box around the plot.
}}

\item{lp_list}{a list of parameters for lines and points.
\itemize{
  \item \code{type_ell}: type of ellipsoid visualization (e.g., "wire").
  \item \code{col_ell}: ellipsoid color.
  \item \code{alpha_ell}: alpha for ellipsoid color.
  \item \code{size_p}: size of points.
  \item \code{col_p}: point color.
  \item \code{alpha_p}: alpha for point color.
}}
}
\value{
A 3D plot displaying ellipsoid projections.
}
\description{
Plots all 3D combinations of variables from niche features
or a background dataset.
}
