% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_evniche.R
\name{run_evniche}
\alias{run_evniche}
\title{Run \emph{evniche}}
\usage{
run_evniche(in_browser = TRUE, ...)
}
\arguments{
\item{in_browser}{(logical) whether to run \code{\link{evniche}} in browser.
If TRUE evniche will run in browser (when defined in\code{[base]{options}}).
If FALSE, the app will run as defined by default in shiny.}

\item{...}{additional arguments to be passed to \code{\link[shiny]{runApp}}.}
}
\description{
Run the \emph{evniche} application.
}
\examples{
if (interactive()){
  run_evniche()
}
}
