% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_covar_helpers.R
\name{var_from_range}
\alias{var_from_range}
\title{Variance from range}
\usage{
var_from_range(range)
}
\arguments{
\item{range}{matrix of two rows (minimum and maximum) x as many columns as
variables to consider.}
}
\value{
vector of variance values for all variables considered.
}
\description{
Variance from range
}
