% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_data.R
\name{virtual_data}
\alias{virtual_data}
\title{Generate new data based on virtual niche}
\usage{
virtual_data(features, from = c("ellipsoid", "prediction"),
             data = NULL, prediction = NULL, n = 100, tol = 1e-8)
}
\arguments{
\item{features}{list of characteristics that defined the ellipsoid. This
list ca be obtained using the function \code{\link{ell_features}}.}

\item{from}{(character) where to generate or sample data from. Options are
"ellipsoid" or "prediction". Default = "ellipsoid" .}

\item{data}{matrix or data.frame containing values (at least environmental
values) used to obtain \code{prediction} based on \code{features}.
Needed if \code{from} = "prediction" and the list used in argument
\code{prediction} contains results of class numeric. Not required if the list
used in argument \code{prediction} contains results of class RasterLayer.}

\item{prediction}{list of predictions based on \code{features} and
\code{data}, obtained using the function \code{\link{ell_predict}}.
Needed if \code{from} = "prediction".}

\item{n}{(numeric) size of data to be generated (number of points).
Default = 100.}

\item{tol}{(numeric) the tolerance for detecting linear dependencies.
Default = 1e-8.}
}
\value{
A matrix or data.frame with the virtual data generated.
}
\description{
Generate new data based on virtual niche
}
\details{
Generation of virtual data is done using the function
\code{\link[MASS]{mvrnorm}} when \code{from} = "ellipsoid".

Virtual data is generated according to suitability values (multivariate
normal probabilities) when \code{from} = "prediction". In this case defining
the arguments \code{data} and \code{prediction} is mandatory if the elements
in \code{prediction} are of class numeric. If elements in \code{prediction}
are of class RasterLayer, \code{data} is not required.
}
