% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_covar_helpers.R
\name{covariance_limits}
\alias{covariance_limits}
\title{Covariance value limits given variable ranges}
\usage{
covariance_limits(range, tol = 1e-8)
}
\arguments{
\item{range}{range values for the variables considered.}

\item{tol}{a value of tolerance for tests. Default = 1e-8.}
}
\value{
a data.frame with estimated minimum and maximum covariance values
for the variables, given the ranges provided.
}
\description{
Covariance value limits given variable ranges
}
